Move and modify sort pipe for multi-use
