+++ src/app/app.module.ts
@@ -11,8 +11,8 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { FilterPipe } from './shared/filter.pipe';
 import { Globals } from './shared/app.global';
-import { ReversePipe } from './../pipe_manipulation';
-import { SortPipe } from './../pipe_manipulation';
+import { ReversePipe } from './shared/pipe_manipulation';
+import { SortPipe } from './shared/pipe_manipulation';
 import { UserSelectorComponent } from './user-selector/user-selector.component';
 
 import { CollegesComponent } from './colleges/colleges.component';
