+++ src/app/shared/pipe_manipulation.ts
